Moved ParserException into decodex.data.exception package.
