+++ src/main/java/decodex/parser/Parser.java
@@ -4,7 +4,7 @@ import decodex.commands.Command;
 import decodex.commands.ExitCommand;
 import decodex.data.DataManager;
 import decodex.modules.ModuleManager;
-import decodex.parser.exception.ParserException;
+import decodex.data.exception.ParserException;
 import decodex.ui.Ui;
 
 import java.util.Arrays;
